requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.2
idna==3.10
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
slack_sdk==3.36.0
urllib3==2.5.0
